add condition, pop up alert

+++ check_pixel.php
@@ -28,12 +28,22 @@ foreach ($xmlimages as $xmlimage) {
 }
 fwrite($fp, "zip -r download.zip download");
 fclose($fp);
-exec("chmod +x url.file");
-exec("./url.file ");
-exec("rm -rf uploads/* | rm -rf url.file | rm -rf download/");
-echo '<h4><a href="http://172.16.0.20/xml_image/download.zip" target="_blank">Download all image in zip</h4>';
+exec("cat url.file | grep wget",$hasil);
+if(!$hasil)
+{
+	echo "<h3>No Image Found in XML File</h3>";
+}
+else
+	{
+	exec("chmod +x url.file");
+    exec("./url.file ");
+    exec("rm -rf uploads/* | rm -rf url.file | rm -rf download/");
+	echo "<script>alert('Download Completed');</script>";
+    echo '<h3><a href="http://172.16.0.20/xml_image/download.zip" target="_blank">Download all image in zip</h3>';
+	}
 //---------------------------------------------------------------------------------------//
 
 ?>
+
 </body>
 </html>
